telephony: Set accept_ra==2 for rmnet_usbx interfaces
When ipv6 forwarding is active, rmnet_usbX interfaces
are not receiving an ipv6 address. This will happen for
instance, when tethering is enabled before the rmnet_usbX
interface is setup.

Set accep_ra==2, for those interfaces, which allows them
to act on router advertisements even when forwarding
is active

CRs-Fixed: 377325
Change-Id: I6bcbc39f845309e52793b2f87eb4f529c320bd94