system/core: usb : Provide usb composition support for 8996/8096
Include cases for 8096/8996 in two usb composition scripts that
enable adb and diag. Update the target script to use soC's machine
property instead of build_id property. The build_id property is not
set for 8096/8996 and will result in incorrect behavior. The machine
property provides the expected output and supports many devices as is.

Change-Id: I0366faf91189b42d1bfc4b616e6759d6ce0e0fe5
Signed-off-by: Clyde Byrd III <cbyrd@codeaurora.org>